requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.1.0
pydantic==2.11.4
email-validator==2.3.0
fastapi-login==1.10.3
passlib[argon2]==1.7.4
python-multipart==0.0.20
cloudinary==1.44.1
